+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop to control caret alignment inside `GroupOption`: `caretAlign`
+- Prop to adjust height of group option on mobile.
 
 ## [0.1.1] - 2020-05-08